requirements.txt
streamlit
google-genai
pandas
gTTS